# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Swagger Finanting
  description: This is a sample server Finanting server.
  contact:
    email: jeanbaptiste.wittner@outlook.com
  license:
    name: GPL-3.0 License
    url: https://www.gnu.org/licenses/gpl-3.0.html
  version: 1.0.5
...
servers:
  - url: http://localhost:9080
...
paths:
  /authentication/login:
    post:
      tags:
        - authentication
      summary: Login
      operationId: login
      requestBody:
        description: Object that need to be authenticated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginParameter'
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginDTO'
...
  /category:
    put:
      tags:
        - category
      summary: aaaaaaaaaaaaa
      operationId: createCategory
      requestBody:
        description: aaaaaaaaaaaaa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryParameter'
      responses:
        201:
          description: aaaaaaaaaaaaa
    get:
      tags:
        - category
      summary: aaaaaaaaaaaaa
      operationId: getUserCategories
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreeCategoryDTO'
  /category/{categoryId}:
    delete:
      tags:
        - category
      summary: aaaaaaaaaaaaa
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: path
          description: aaaaaaaaaaaaa
          required: true
          schema:
            type: integer
      responses:
        200:
          description: aaaaaaaaaaaaa
    post:
      tags:
        - category
      summary: aaaaaaaaaaaaa
      operationId: updateCategory
      parameters:
        - name: categoryId
          in: path
          description: aaaaaaaaaaaaa
          required: true
          schema:
            type: integer

      requestBody:
        description: aaaaaaaaaaaaa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryParameter'
      responses:
        200:
          description: aaaaaaaaaaaaa
  /category/{groupId}:
    get:
      tags:
        - category
      summary: aaaaaaaaaaaaa
      operationId: getGroupCategories
      parameters:
        - name: groupId
          in: path
          description: aaaaaaaaaaaaa
          required: true
          schema:
            type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreeCategoryDTO'
...
components:
  securitySchemes:
    jwtAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginDTO:
      type: object
      properties:
        jwt:
          type: string
      required:
        - jwt
...
    TreeCategoryDTO:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        abbreviation:
          type: string
        description:
          type: string
        categoryType:
          type: string
          enum:
            - EXPENSE
            - REVENUE
        childTreeCategoriesDTOs:
          type: array
          items:
            $ref: '#/components/schemas/TreeCategoryDTO'
      required:
      - id
      - label
      - abbreviation
      - categoryType
      - childTreeCategoriesDTOs
...
    LoginParameter:
      type: object
      properties:
        userName:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
...
    CategoryParameter:
      type: object
      properties:
        parentId:
          type: integer
        label:
          type: string
        abbreviation:
          type: string
          minLength: 3
          maxLength: 6
        description:
          type: string
        groupName:
          type: string
        categoryType:
          type: string
          enum:
            - EXPENSE
            - REVENUE
      required:
        - label
        - abbreviation
    UpdateCategoryParameter:
      type: object
      properties:
        parentId:
          type: integer
        label:
          type: string
        abbreviation:
          type: string
          minLength: 3
          maxLength: 6
        description:
          type: string
        categoryType:
          type: string
          enum:
            - EXPENSE
            - REVENUE
      required:
        - label
        - abbreviation
...